requirements.txt
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.40
scikit-learn>=1.3.0
torch>=2.1.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
pyyaml>=6.0.0
tabulate>=0.9.0
etsformer-pytorch>=0.1.7
